pyproject.toml
[tool.poetry]
name = "implementing-webhooks"
version = "0.1.0"
description = ""
authors = ["Brian Ibbotson <brian@quinceleaf.dev>"]

[tool.poetry.dependencies]
python = "^3.8"
Django = "^3.2"
psycopg2-binary = "^2.8.6"
gunicorn = "^20.1.0"
django-redis = "^4.12.1"
django-filter = "^2.4.0"
django-simple-history = "^2.12.0"
djangorestframework-simplejwt = "^4.7.0"
django-cors-headers = "^3.7.0"
PyJWT = "^2.1.0"
python-dotenv = "^0.15.0"
pytz = "^2021.1"
django-debug-toolbar = "^3.2"
coverage = "^5.5"
python-ulid = "^1.0.3"

[tool.poetry.dev-dependencies]
black = "^20.8b1"
Faker = "^6.6.0"
factory-boy = "^3.2.0"
flake8 = "^3.8.4"
model-bakery = "^1.2.1"
selenium = "^3.141.0"
PyNaCl = "^1.4.0"
requests = "^2.25.1"

[build-system]
requires = ["poetry-core==1.0.0","setuptools>=50", "cryptography<=3.3.2"]
build-backend = "poetry.core.masonry.api"
